completed react hot button component exercise
